@@ -11,7 +11,7 @@ dependencies:
     sdk: flutter
   cupertino_icons: ^1.0.5
 
-  youtube_player_flutter: ^8.1.2
+  youtube_player_flutter: ^9.1.3
 
   video_player: ^2.8.2
 
@@ -21,9 +21,9 @@ dependencies:
   
   image_picker: ^1.0.7
   
-  file_picker: ^6.1.1
+  file_picker: ^10.3.7
 
-  confetti: ^0.7.0
+  confetti: ^0.8.0
 
   # HTTP + авторизация
   dio: ^5.0.0
@@ -41,13 +41,22 @@ dependencies:
 
   json_annotation: ^4.8.0
   url_launcher: ^6.3.2
+  
+  # Для записи аудио
+  record: ^6.1.2
+  
+  # Для воспроизведения аудио
+  audioplayers: ^6.5.1
+  
+  # Для работы с разрешениями
+  permission_handler: ^12.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   build_runner: ^2.5.0
   json_serializable: ^6.6.0
-  flutter_lints: ^3.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
